Extract to library
Extract functionality into other modules to simplify the main
application.
