+++ Mirin.cabal
@@ -13,10 +13,12 @@ build-type:          Simple
 extra-source-files:  README.md
 cabal-version:       >=1.10
 
-executable Mirin
-  hs-source-dirs:      app
-  main-is:             Main.hs
+library
+  hs-source-dirs:      src
   default-language:    Haskell2010
+  exposed-modules:     Util
+                     , Settings
+                     , Model
   build-depends:       base >= 4.7 && < 5
                      , Spock >= 0.11
                      , mtl
@@ -31,3 +33,15 @@ executable Mirin
                      , http-types
                      , wai
                      , resourcet
+                     , resource-pool
+
+
+executable Mirin
+  hs-source-dirs:      app
+  main-is:             Main.hs
+  default-language:    Haskell2010
+  build-depends:       base
+                     , Mirin
+                     , Spock
+                     , persistent-mysql
+                     , transformers
